requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
blinker==1.8.2
cattrs==23.2.3
channels==4.1.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.1
dash==2.9.3
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==5.0.6
django-bootstrap-datepicker-plus==5.0.5
django-chartjs==2.3.0
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-plotly-dash==2.3.1
django-schema-graph==3.1.0
dpd_components==0.1.0
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
Flask==3.0.3
graphviz==0.20.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pep8-naming==0.13.2
phonenumbers==8.13.37
platformdirs==4.2.1
plotly==5.22.0
pycodestyle==2.9.1
pydantic==2.7.1
pydantic_core==2.18.2
pydotplus==2.0.2
pyflakes==2.5.0
pyparsing==3.1.2
sqlparse==0.5.0
tenacity==8.3.0
typing_extensions==4.11.0
tzdata==2024.1
Werkzeug==3.0.3
